fix: layout head was not rendered
