+++ apps/web/src/app/head.tsx
@@ -1,0 +1,5 @@
+import { Head } from '@mss/web/components/Head'
+
+const defaultExport = Head
+
+export default defaultExport
