+++ apps/web/src/components/Head/index.ts
@@ -1,0 +1,1 @@
+export { Head } from './Head'
